move constants and fix aliases
