+++ .vscode/settings.json
@@ -9,16 +9,12 @@
 		"typescript", // Enable .ts
 		"typescriptreact" // Enable .tsx
 	],
-	"typescript.tsdk": "node_modules/typescript/lib",
 	"[choicescript]": {
 		"editor.quickSuggestions": {
 			"comments": false,
 			"strings": false,
 			"other": false
 		}
-	},
-	"debug.javascript.defaultRuntimeExecutable": {
-		"pwa-node": "/Users/echo/.local/share/mise/shims/node"
 	},
 	"editor.inlineSuggest.enabled": false
 }
