+++ eslint.config.mjs
@@ -9,12 +9,16 @@ import tseslint from 'typescript-eslint';
 const eslintConfig = tseslint.config(
 	eslint.configs.recommended,
 	tseslint.configs.recommended,
-	perfectionist.configs['recommended-alphabetical'],
+	perfectionist.configs['recommended-natural'],
 	eslintPluginAstro.configs['jsx-a11y-recommended'],
 	eslintConfigPrettier,
 	{
 		rules: {
 			'astro/jsx-a11y/media-has-caption': ['off'],
+			'perfectionist/sort-imports': [
+				'warn',
+				{ newlinesBetween: true, type: 'natural' },
+			],
 		},
 	},
 	{
