+++ src/components/Footer.astro
@@ -1,5 +1,5 @@
 ---
-import { SITE_TITLE } from '~/consts';
+import { SITE_TITLE } from '~/lib/constants';
 
 import Link from './Link.astro';
