+++ src/components/Header.astro
@@ -1,5 +1,5 @@
 ---
-import { SITE_DESCRIPTION, SITE_TITLE } from '~/consts';
+import { SITE_DESCRIPTION, SITE_TITLE } from '~/lib/constants';
 
 import HeaderLink from './HeaderLink.astro';
 import Link from './Link.astro';
