+++ src/layouts/BaseLayout.astro
@@ -5,7 +5,7 @@ import BaseHead from '~/components/BaseHead.astro';
 import Footer from '~/components/Footer.astro';
 import Header from '~/components/Header.astro';
 import StarField from '~/components/StarField.astro';
-import { SITE_DESCRIPTION, SITE_TITLE } from '~/consts';
+import { SITE_DESCRIPTION, SITE_TITLE } from '~/lib/constants';
 
 interface Props extends HTMLAttributes<'main'> {
 	class?: string;
